add(print_packet error): automatically detect errors
