+++ inc/ft_ping.h
@@ -229,9 +229,12 @@ t_list      *session_init_all(uint16_t pid, const t_list *hosts, const t_arg_d *
 // void     session_print_sum(t_sum *session);
 // void     session_end(t_list **sessions);
 
-uint16_t    packet_checksum_calculate(const char *buffer, size_t size);
 int         packet_send(int sockfd, const t_host *dest, const t_packet *packet);
 int         packet_receive(int sockfd, t_seq *sequence);
+void        packet_print(const t_seq *sequence, float time_enlapsed);
+void        packet_cpy(t_packet *target, const t_packet *source);
+uint16_t    packet_checksum_calculate(const char *buffer, size_t size);
+int         packet_verify_headers(const t_seq *sequence, uint8_t type, uint8_t code);
 
 void        host_lookup(t_host *host, char *raw_addr, bool do_resolution);
 void        host_get_ip(struct sockaddr *addr_buff);
